+++ src/asphalt/type.clj
@@ -3,6 +3,11 @@
     [java.sql PreparedStatement ResultSet]))
 
 
+(defprotocol IConnectionSource
+  (obtain-connection            [this] "Obtain connection from the source")
+  (return-connection [this connection] "Return connection to the source"))
+
+
 (defprotocol ISql
   (get-sql    [this] "Return SQL string to be executed")
   (set-params [this ^PreparedStatement prepared-statement params] "Set prepared-statement params")